requirements.txt
pyodbc==5.2.0
fastapi==0.115.6
pytz==2024.2
uvicorn==0.34.0
pandas==2.2.3
pyjwt==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
SQLAlchemy==1.4.26
python-dateutil==2.8.2
pydantic==2.10.5
paho-mqtt==1.6.1
requests==2.32.3
redis==5.2.1
requests==2.32.3
streamlit==1.44.1
streamlit-cookies-controller==0.0.4
plotly==6.0.1
matplotlib==3.10.3
tqdm==4.67.1
scipy==1.16.0
sko==0.5.7
seaborn==0.13.2
cvxpy==1.7.0
deap==1.4.3
networkx==3.5
pyinform==0.2.0
ccxt==4.4.94
python-telegram-bot==22.3